requirements.txt
PyTelegramBotApi
psycopg2-binary
telebot